requirements.txt
PyYAML
matplotlib
ipywidgets
ipykernel
setuptools
langchain
langchain-community
langchain-experimental
langchain-openai
langchain-ollama
langgraph

langchain-azure-ai[opentelemetry]
autogen
autogen_core
